+++ AIRSIS/addMissingElevation.R
@@ -1,28 +1,14 @@
 # Use a web service to add missing elevation data AFTER updating with OpenCage
 
-# library(MazamaLocationUtils)
-#
-# collectionName <- "airsis_PM2.5_sites_1000"
-#
-# library(MazamaLocationUtils)
-#
-# collectionDir <- "AIRSIS"
-# collectionName <- "airsis_PM2.5_sites_1000"
-# collectionFile <- paste0(collectionName, ".rda")
-#
-# # ----- Load known locations ---------------------------------------------------
-#
-# download.file(
-#   file.path("http://data-monitoring_v2-c1.airfire.org/monitoring-v2/known-locations", collectionFile),
-#   destfile = file.path(".", collectionDir, collectionFile)
-# )
-#
-# setLocationDataDir(file.path(".", collectionDir))
-#
-# locationTbl <- table_load(collectionName)
+library(MazamaLocationUtils)
 
+collectionName <- "airsis_PM2.5_sites_1000"
 
-# ----- Review missing elevations ----------------------------------------------
+# ----- Load and Review --------------------------------------------------------
+
+setLocationDataDir("./AIRSIS")
+
+locationTbl <- table_load(collectionName)
 
 missingElevation_mask <- is.na(locationTbl$elevation)
 
@@ -31,7 +17,8 @@ locationTbl[missingElevation_mask,] %>%
 
 # ----- Add elevations ---------------------------------------------------------
 
-message(sprintf("Getting %d missing elevations ...", sum(missingElevation_mask)))
+missingCount <- sum(missingElevation_mask)
+message(sprintf("Getting %d missing elevations ...", missingCount))
 
 # NOTE:  This is a manual operation so print out progress
 count <- 0
@@ -39,7 +26,7 @@ for ( i in which(missingElevation_mask) ) {
 
   count <- count + 1
   if ( (count %% 10) == 0 )
-    message(sprintf("Working on %d/%d ...", count, length(missingElevation_mask)))
+    message(sprintf("Working on %d/%d ...", count, missingCount))
 
   result <- try({
 
@@ -61,7 +48,6 @@ for ( i in which(missingElevation_mask) ) {
 
 locationTbl[missingElevation_mask,] %>%
   MazamaLocationUtils::table_leaflet(extraVars = c("elevation", "address"))
-
 
 # ----- Save the table ---------------------------------------------------------